requirements_test.txt
flake8-docstrings
flake8
pydocstyle
pytest
vcrpy
